@@ -1,13 +1,18 @@
 {
     "name": "yakamara/redaxo_yrewrite",
 
+    "require": {
+        "php": ">=8.1"
+    },
+
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "^3.4",
-        "friendsofredaxo/linter": "^1.3",
-        "redaxo/php-cs-fixer-config": "^1.0",
-        "redaxo/psalm-plugin": "^1.0",
-        "redaxo/source": "^5.8",
-        "vimeo/psalm": "^4.25",
+        "friendsofphp/php-cs-fixer": "^3.15.1",
+        "friendsofredaxo/linter": "^1.4",
+        "j13k/yaml-lint": "@dev",
+        "redaxo/php-cs-fixer-config": "^2.1.1",
+        "redaxo/psalm-plugin": "^2.0",
+        "redaxo/source": "^5.15",
+        "vimeo/psalm": "^5.8.0",
         "yakamara/redaxo_yform": "dev-master"
     },
 
@@ -19,17 +24,15 @@
 
     "scripts": {
         "cs": "php-cs-fixer fix -v --ansi",
-        "psalm": "psalm --no-suggestions"
+        "psalm": "psalm --no-suggestions",
+        "post-update-cmd": "@composer bump"
     },
 
     "config": {
         "optimize-autoloader": true,
         "sort-packages": true,
         "platform": {
-            "php": "7.3.0"
-        },
-        "allow-plugins": {
-            "composer/package-versions-deprecated": true
+            "php": "8.1.16"
         }
     }
 }
